# docs/openapi.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "glance API",
    "version": "0.3.0-dev",
    "description": "Public HTTP API of glance, the lean Plane-like issue tracker (Go monolith, single binary). All paths are relative to the server root. NOTE ON FORMAT: this file is written in JSON syntax, which is valid YAML 1.2 — docs/openapi.yaml and the served GET /api/v1/openapi.json are the SAME single file with no build-time conversion; keep it JSON-compatible so both parsers accept it."
  },
  "servers": [
    { "url": "/", "description": "Same origin as the glance server (self-hosted)" }
  ],
  "security": [],
...
  "paths": {
...
    "/api/v1/workspaces/{slug}/import": {
      "post": {
        "tags": ["workspaces"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Import a workspace data archive",
        "description": "Admin only. Restores a glance-export/1 JSON archive (as produced by GET /api/v1/workspaces/{slug}/export) into this workspace. multipart/form-data with a `file` field (the archive JSON); 100MB limit. Restore-by-UUID: rows whose UUID already exists are skipped, so re-import is idempotent. Members are matched to instance users by email (never auto-created); unmapped members are skipped and reported. Attachment binaries are never exported and cannot be restored — attachment metadata is counted as skipped. Returns an honest import report {format, source_workspace, target_workspace, imported, skipped}. Archives with an unknown format version are refused with 400 and import nothing.",
        "parameters": [ { "$ref": "#/components/parameters/Slug" } ],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "required": ["file"], "properties": { "file": { "type": "string", "format": "binary", "description": "glance-export/1 archive JSON" } } } } } },
        "responses": { "200": { "description": "Import report", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArchiveImportReport" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "413": { "$ref": "#/components/responses/PayloadTooLarge" } }
      }
    },
...
    "/api/v1/workspaces/{slug}/projects/{identifier}/imports": {
      "post": {
        "tags": ["import"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Import issues from CSV",
        "description": "multipart/form-data with a `file` field (the CSV) and a `mapping` field (JSON column mapping). 10MB limit. Valid rows are inserted; per-row failures are reported.",
        "parameters": [ { "$ref": "#/components/parameters/Slug" }, { "$ref": "#/components/parameters/Identifier" } ],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "required": ["file", "mapping"], "properties": { "file": { "type": "string", "format": "binary" }, "mapping": { "type": "string", "description": "JSON import mapping (see ImportMapping)" } } } } } },
        "responses": { "200": { "description": "Import result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportResult" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "413": { "$ref": "#/components/responses/PayloadTooLarge" } }
      }
    },
    "/api/v1/workspaces/{slug}/projects/{identifier}/imports/preview": {
      "post": {
        "tags": ["import"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Preview a CSV import",
        "description": "Parses and validates the multipart CSV + mapping without inserting anything.",
        "parameters": [ { "$ref": "#/components/parameters/Slug" }, { "$ref": "#/components/parameters/Identifier" } ],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "required": ["file", "mapping"], "properties": { "file": { "type": "string", "format": "binary" }, "mapping": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Preview result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportResult" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "413": { "$ref": "#/components/responses/PayloadTooLarge" } }
      }
    },
    "/api/v1/workspaces/{slug}/projects/{identifier}/imports/template.csv": {
      "get": {
        "tags": ["import"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Download the CSV import template",
        "parameters": [ { "$ref": "#/components/parameters/Slug" }, { "$ref": "#/components/parameters/Identifier" } ],
        "responses": { "200": { "description": "CSV template", "content": { "text/csv": { "schema": { "type": "string" } } } } }
      }
    },
    "/api/v1/workspaces/{slug}/projects/{identifier}/imports/github/preview": {
      "post": {
        "tags": ["import"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Preview a GitHub issues import",
        "description": "Fetches issues from the GitHub REST API (paginated, pull requests excluded) and reports how the first 10 would map: labels (missing ones would be created), assignee matches, comment counts, milestones (skipped). No writes. The token travels in the body only and is never stored. Member (15)+.",
        "parameters": [ { "$ref": "#/components/parameters/Slug" }, { "$ref": "#/components/parameters/Identifier" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["owner", "repo"], "properties": { "owner": { "type": "string" }, "repo": { "type": "string" }, "token": { "type": "string", "description": "GitHub personal access token; empty works for public repos within unauthenticated rate limits" }, "state_filter": { "type": "string", "enum": ["open", "closed", "all"], "default": "open" }, "max": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } } } } } },
        "responses": { "200": { "description": "Preview result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GitHubImportPreview" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "429": { "$ref": "#/components/responses/RateLimited" }, "502": { "$ref": "#/components/responses/BadGateway" } }
      }
    },
    "/api/v1/workspaces/{slug}/projects/{identifier}/imports/github": {
      "post": {
        "tags": ["import"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Import issues from GitHub",
        "description": "Imports GitHub issues as glance issues: labels are created when missing, comments are imported with attribution, assignees are matched best-effort by public email, milestones are skipped (glance has none). Idempotent via an imported-from marker in each issue's description: re-running skips already-imported issues. The token travels in the body only and is never stored. Member (15)+.",
        "parameters": [ { "$ref": "#/components/parameters/Slug" }, { "$ref": "#/components/parameters/Identifier" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["owner", "repo"], "properties": { "owner": { "type": "string" }, "repo": { "type": "string" }, "token": { "type": "string", "description": "GitHub personal access token; empty works for public repos within unauthenticated rate limits" }, "state_filter": { "type": "string", "enum": ["open", "closed", "all"], "default": "open" }, "max": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } } } } } },
        "responses": { "200": { "description": "Import result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GitHubImportResult" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "429": { "$ref": "#/components/responses/RateLimited" }, "502": { "$ref": "#/components/responses/BadGateway" } }
      }
    },
    "/api/v1/workspaces/{slug}/projects/{identifier}/imports/jira/preview": {
      "post": {
        "tags": ["import"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Preview a Jira Cloud issues import",
        "description": "Fetches issues from the Jira Cloud REST v3 API (POST /rest/api/3/search/jql, token pagination) and reports how the first 10 would map: status → state (unmatched statuses would create new states), labels (missing ones would be created), priority, assignee matches, comment counts. No writes. Jira Cloud only — site is the Atlassian subdomain (host pinned to <site>.atlassian.net); Server/Data Center are not supported. The API token travels in the body only and is never stored. Member (15)+.",
        "parameters": [ { "$ref": "#/components/parameters/Slug" }, { "$ref": "#/components/parameters/Identifier" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["site", "email", "api_token", "project_key"], "properties": { "site": { "type": "string", "description": "Atlassian subdomain only, e.g. \"acme\" (→ acme.atlassian.net)" }, "email": { "type": "string" }, "api_token": { "type": "string", "description": "Jira Cloud API token (from id.atlassian.com); sent as HTTP Basic auth, never stored" }, "project_key": { "type": "string", "description": "Jira project key, e.g. \"PROJ\"" }, "max": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } } } } } },
        "responses": { "200": { "description": "Preview result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JiraImportPreview" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "429": { "$ref": "#/components/responses/RateLimited" }, "502": { "$ref": "#/components/responses/BadGateway" } }
      }
    },
    "/api/v1/workspaces/{slug}/projects/{identifier}/imports/jira": {
      "post": {
        "tags": ["import"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Import issues from Jira Cloud",
        "description": "Imports Jira Cloud issues as glance issues: statuses map to states by name (unmatched statuses create new states in the unstarted group), labels are created when missing, priorities are mapped, comments are imported with attribution and original timestamps, assignees are matched best-effort by email, sprints are skipped (glance cycles are manual), subtasks/epics import as flat issues. Idempotent via an imported-from marker in each issue's description: re-running skips already-imported issues. Jira Cloud only. The API token travels in the body only and is never stored. Member (15)+.",
        "parameters": [ { "$ref": "#/components/parameters/Slug" }, { "$ref": "#/components/parameters/Identifier" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["site", "email", "api_token", "project_key"], "properties": { "site": { "type": "string", "description": "Atlassian subdomain only, e.g. \"acme\" (→ acme.atlassian.net)" }, "email": { "type": "string" }, "api_token": { "type": "string", "description": "Jira Cloud API token (from id.atlassian.com); sent as HTTP Basic auth, never stored" }, "project_key": { "type": "string", "description": "Jira project key, e.g. \"PROJ\"" }, "max": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } } } } } },
        "responses": { "200": { "description": "Import result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JiraImportResult" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "429": { "$ref": "#/components/responses/RateLimited" }, "502": { "$ref": "#/components/responses/BadGateway" } }
      }
    },
    "/api/v1/workspaces/{slug}/projects/{identifier}/imports/trello/preview": {
      "post": {
        "tags": ["import"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Preview a Trello board import",
        "description": "Fetches cards from the Trello REST v1 API (before-cursor pagination) and reports how the first 10 would map: list → state (unmatched lists would create new states), labels (missing ones would be created, with Trello palette colors), due date → target date, assignee matches, comment/checklist badge counts (comment bodies are NOT fetched at preview time). No writes. Trello Cloud only — host pinned to https://api.trello.com. Checklists are skipped (glance has no checklist model) and their counts are reported. The API key + token travel in the body only and are never stored. Member (15)+.",
        "parameters": [ { "$ref": "#/components/parameters/Slug" }, { "$ref": "#/components/parameters/Identifier" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["api_key", "board_id"], "properties": { "api_key": { "type": "string", "description": "Trello API key (from trello.com/app-key); sent as a key query param to api.trello.com only, never stored" }, "token": { "type": "string", "description": "Trello token; required for private boards, empty works for public boards" }, "board_id": { "type": "string", "description": "24-char board id, 8-char short link, or an https trello.com board URL" }, "max": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } } } } } },
        "responses": { "200": { "description": "Preview result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TrelloImportPreview" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "429": { "$ref": "#/components/responses/RateLimited" }, "502": { "$ref": "#/components/responses/BadGateway" } }
      }
    },
    "/api/v1/workspaces/{slug}/projects/{identifier}/imports/trello": {
      "post": {
        "tags": ["import"], "security": [{ "sessionCookie": [] }, { "apiToken": [] }],
        "summary": "Import cards from a Trello board",
        "description": "Imports Trello cards as glance issues: lists map to states by name (unmatched lists create new states in the unstarted group), labels are created when missing (Trello palette colors become label colors, color-only labels become the capitalized color name), card desc becomes the description (Markdown kept), due date becomes the target date, comments are imported with attribution and original timestamps, assignees are matched best-effort by email, archived cards are excluded (counted as archived_skipped), checklists are skipped and counted (glance has no checklist model). Idempotent via an imported-from marker in each issue's description: re-running skips already-imported cards. Trello Cloud only. The API key + token travel in the body only and are never stored. Member (15)+.",
        "parameters": [ { "$ref": "#/components/parameters/Slug" }, { "$ref": "#/components/parameters/Identifier" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["api_key", "board_id"], "properties": { "api_key": { "type": "string", "description": "Trello API key (from trello.com/app-key); sent as a key query param to api.trello.com only, never stored" }, "token": { "type": "string", "description": "Trello token; required for private boards, empty works for public boards" }, "board_id": { "type": "string", "description": "24-char board id, 8-char short link, or an https trello.com board URL" }, "max": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } } } } } },
        "responses": { "200": { "description": "Import result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TrelloImportResult" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "429": { "$ref": "#/components/responses/RateLimited" }, "502": { "$ref": "#/components/responses/BadGateway" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey", "in": "cookie", "name": "glance_session",
        "description": "Session cookie set by POST /api/v1/auth/otp/verify or the OAuth callback. HttpOnly."
      },
      "apiToken": {
        "type": "http", "scheme": "bearer",
        "description": "Personal API token minted via POST /api/v1/tokens (prefix gl_). Bearer <redacted> that do not start with gl_ are ignored by the middleware. Scopes: read (GETs), write (implies read)."
      }
    },
    "parameters": {
      "Slug": { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Workspace slug" },
      "Identifier": { "name": "identifier", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Project identifier (uppercased short key)" },
...
    },
    "responses": {
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Missing or invalid auth", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Authenticated but not allowed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "RateLimited": { "description": "Too many requests", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "PayloadTooLarge": { "description": "Request body exceeds the size limit", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "BadGateway": { "description": "Upstream service error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
...
      "GitHubImportPreview": {
        "type": "object",
        "properties": {
          "source": { "type": "string" },
          "total": { "type": "integer" },
          "rows": { "type": "array", "items": { "type": "object", "properties": {
            "number": { "type": "integer" },
            "title": { "type": "string" },
            "state": { "type": "string" },
            "labels": { "type": "array", "items": { "type": "string" } },
            "new_labels": { "type": "array", "items": { "type": "string" } },
            "assignees": { "type": "array", "items": { "type": "string" } },
            "assignee_matched": { "type": "boolean" },
            "comments": { "type": "integer" },
            "milestone": { "type": "string", "nullable": true },
            "already_imported": { "type": "boolean" }
          } } }
        }
      },
...
    }
  }
}